Streamlit pages cleaned up, and working in testing
